requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch>=2.4.1
pillow>=10.3.0
keybert
spacy
sentencepiece
